Приложение отдебажено. Изначально был прописан неправильный запрос к API. Так же в LogCat отследил неверный тип данных в автоматически сгенерированных классах. В итоге все типы данных были заменены. Приложение работает

+++ app/src/main/java/com/example/courseworknewversionlib/network/RetailPrice.kt
@@ -5,7 +5,7 @@ import com.google.gson.annotations.SerializedName
 
 data class RetailPrice (
 
-  @SerializedName("amountInMicros" ) var amountInMicros : Int?    = null,
+  @SerializedName("amountInMicros" ) var amountInMicros : Long?    = null,
   @SerializedName("currencyCode"   ) var currencyCode   : String? = null
 
 )
